requirements.txt
# Core API Framework
# Uncomment based on your choice:
# fastapi==0.104.1
# flask==3.0.0

# Configuration & YAML
pyyaml==6.0.1

# HTTP Client
requests==2.31.0

# Type checking & validation (compatible with Python 3.14)
pydantic>=2.5.0

# Development tools
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
ruff>=0.1.8
mypy>=1.7.0

# Logging (Optional)
# python-logging-loki==0.3.1  # Uncomment if you need structured logging support
# Available versions: 0.3.1, 0.3.0, 0.2.0, 0.1.0

# Environment variables
python-dotenv==1.0.0
